move from de-CH to "neutral" de-DE language
The content itself is currently targeting de-DE because of the ß char, so it would made sense to move the culture file to de-DE.

Besides that: Most systems only allow the "main" language as a UI language (e.g. you only can set Windows to the de-DE UiCulture or en-US UiCulture).